+++ src/elite/perf/EliteSharpLib.Benchmarks/PlanetBenchmarks.cs
@@ -31,7 +31,7 @@ public class PlanetBenchmarks : IDisposable
         Useful.Abstraction.ScreenManager<Views.Screen, Views.IView> views = new(keyboard);
         GameState gameState = new(views);
         _graphics = SoftwareGraphics.Create(ScreenWidth, ScreenHeight, (_) => { }, assetLocator);
-        ZBufferRenderer shipRenderer = new(gameState, _graphics, assetLocator);
+        ZBufferRenderer shipRenderer = new(_graphics);
         EliteDraw draw = new(gameState, _graphics, assetLocator, shipRenderer);
         _wireframePlanet = new(draw);
         _solidPlanet = new(draw);